requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.1.0
cryptography==2.9.2
Flask==1.0.2
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
mysqlclient==1.3.12
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pycryptodome==3.9.7
PyMySQL==0.9.3
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.8.0
PyTestReport==0.2.1
pytz==2020.1
PyYAML==3.12
requests==2.22.0
requests-toolbelt==0.9.1
selenium==2.53.6
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0